rename: StartButton을 common components로 이동
동영상 화면에서도 사용하기 위함
