+++ FE/src/components/common/StartButton/StartButton.tsx
