+++ FE/src/page/LandingPage/index.tsx
@@ -1,5 +1,5 @@
 import LandingPageLayout from '@/components/landingPage/LandingPageLayout';
-import StartButton from '@components/landingPage/StartButton';
+import StartButton from '@common/StartButton/StartButton';
 import WelcomeBlurb from '@components/landingPage/WelcomeBlurb';
 import LandingImage from '@components/landingPage/LandingImage';
 import GoogleLoginButton from '@components/landingPage/GoogleLoginButton';
